Implementação do validation e tratando as exceptions personalizadas!

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -19,11 +24,11 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
-
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado!"));
         return new ProductDTO(product);
+
     }
 
     @Transactional(readOnly = true)
@@ -41,20 +46,32 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductDTO update(Long id,ProductDTO productDTO){
-        Product product = productRepository.getReferenceById(id);//Nao vai pro banco mas é monitorado pelo JPA
-        copyDtoToEntity(productDTO,product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+    public ProductDTO update(Long id,ProductDTO productDTO) {
+        try {
+            Product product = productRepository.getReferenceById(id);//Nao vai pro banco mas é monitorado pelo JPA
+            copyDtoToEntity(productDTO, product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
-    public void delete(Long id){
-        productRepository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
 
-    private void copyDtoToEntity(ProductDTO productDTO, Product product) {
+    private void copyDtoToEntity(ProductDTO productDTO,Product product) {
         product.setName(productDTO.getName());
         product.setDescription(product.getDescription());
         product.setPrice(productDTO.getPrice());
